Split the kalman tests into two modules.
